Isolating contents of release note to own file + renaming from 'releaseNotes' to 'releaseNote' to match browser devtools
