+++ postBuildStep.ts
@@ -6,7 +6,7 @@ import path from "path";
 import applyPaddingInlineCssPatch from "./src/host/polyfills/cssPaddingInline";
 import { applyContentSecurityPolicyPatch } from "./src/host/polyfills/inspectorContentPolicy";
 import applyRuntimeImportScriptPathPrefixPatch from "./src/host/polyfills/runtime";
-import {applyAnnouncementNamePatch, applyGithubLinksPatch, applyReleaseNotePatch} from "./src/host/polyfills/releaseNotes";
+import {applyAnnouncementNamePatch, applyGithubLinksPatch, applyReleaseNotePatch} from "./src/host/polyfills/releaseNote";
 import {
     applyAppendTabOverridePatch,
     applyAppendTabConditionsPatch,
