+++ src/host/polyfills/releaseNote.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { getTextFromFile } from "../../test/helpers";
-import * as ReleaseNotePatch from "./releaseNotes";
+import * as ReleaseNotePatch from "./releaseNote";
 
 describe("Release notes text replacement", () => {
     it("verifies that applyReleaseNotePatch correctly replaces what's new content", async () => {
