+++ src/host/polyfills/releaseNoteContent.ts
@@ -1,0 +1,52 @@
+export const ExtensionHighlights = `[
+    {
+        title: ls\`DevTools Extension Console rerouting to VSCode output window\`,
+        subtitle:
+            ls\`The console messages from the DevTools are now rerouted to the VS Code Output window for an integrated debugging experience.\`,
+        link: 'https://github.com/microsoft/vscode-edge-devtools/pull/275',
+    },
+    {
+        title: ls\`Improvements to revealing files in VSCode from the DevTools styles pane\`,
+        subtitle:
+            ls\`Clicking on a property will only navigate when CTRL is held and clicking the styles sheet link will open the file in VS Code.\`,
+        link: 'https://github.com/microsoft/vscode-edge-devtools/pull/264',
+    },
+    {
+        title: ls\`Introducing favicons to the extension sidebar\`,
+        subtitle:
+            ls\`The extension's sidebar will now display the target website's favicon to help organize debugging targets.\`,
+        link: 'https://github.com/microsoft/vscode-edge-devtools/pull/278',
+    },
+    {
+        title: ls\`Changing screencast view cursor from touch to default\`,
+        subtitle:
+            ls\`After receiving feedback, we have changed the type of cursor for the screencast view from a touch cursor to the default cursor to improve precision.\`,
+        link: 'https://github.com/microsoft/vscode-edge-devtools/pull/279',
+    },
+    {
+        title: ls\`Network tool enabled by default\`,
+        subtitle:
+            ls\`The Network tool is now availalbe by default! If you would like to keep it off, uncheck the "Enable Network" checkbox in the VS Code extension settings.\`,
+        link: 'https://github.com/microsoft/vscode-edge-devtools/pull/272',
+    },
+],`
+
+export const DevToolsHighlights = `[
+    {
+        title: ls\`CSS grid overlay improvements and new experimental grid features\`,
+        subtitle:
+            ls\`Grid overlays are now enabled by default with multiple persistent and configurable overlays coming soon.\`,
+        link: 'https://go.microsoft.com/fwlink/?linkid=2142427',
+        },
+        {
+        title: ls\`Highlight all search results in Elements tool\`,
+        subtitle: ls\`Thanks to your feedback, we were able to find and fix a bug in the open-source Chromium project.\`,
+        link: 'https://go.microsoft.com/fwlink/?linkid=2142224'
+        },
+        {
+        title: ls\`Network and Elements panel updates\`,
+        subtitle:
+            ls\`Capture node screenshots shortcut, accessible color suggestion, consistent resource types, and more.\`,
+        link: 'https://go.microsoft.com/fwlink/?linkid=2142153',
+        }
+],`
